+++ client/src/components/Application.scss
