+++ client/src/reducers/application.js
@@ -1,0 +1,18 @@
+export const SET_USERS = 'SET_USERS';
+  
+export const reducer = (state, action) =>{
+
+  const actions ={
+
+    SET_USERS: {
+      ...state,
+      users: action.users
+    }
+
+  }
+
+  if (!actions[action.type]) {
+    throw new Error('Type of action not found');
+  }
+  return actions[action.type];
+}
